Refactor(server): Centralize GEMINI.md discovery logic in server
This is the second step in addressing issue #472 and builds upon
the changes in PR #493.

Moves the hierarchical GEMINI.md file discovery and loading logic
from the CLI package (`packages/cli/src/config/config.ts`) to a new
utility module in the server package
(`packages/server/src/utils/memoryDiscovery.ts`).

- Creates `packages/server/src/utils/memoryDiscovery.ts` with
  functions: `findProjectRoot`, `collectDownwardGeminiFiles`,
  `getGeminiMdFilePathsInternal`, `readGeminiMdFiles`,
  `concatenateInstructions`, and the main exported function
  `loadServerHierarchicalMemory`.
- Updates `packages/cli/src/config/config.ts` to remove the
  above functions and instead calls `loadServerHierarchicalMemory`
  from the `@gemini-code/server` package.
- Modifies `packages/cli/src/ui/hooks/useShowMemoryCommand.ts` to
  no longer display file paths (as this info is not readily
  available after the refactor) but continues to show memory content
  and file count.
- Exports the new memory discovery utility from
  `packages/server/src/index.ts`.
- Addresses related import changes and linting issues.
